Fusion O365 proxy v1.0 (#1)
Initial version of the proxy api.

The api is configured to give fusion test access to f_fusion_test@equinor.com and prod access to f_fusion@equinor.com.

Deployment pipelines have been created and set up to deploy to `fusion-o365-proxy-test` namespace in the test cluster and `fusion-o365-proxy-prod` namespace in the prod cluster.

A separate pipeline has been set up to manage and rotate the client secret used to authenticate with azure ad. The pipeline is on a monthly schedule and will create new secrets with a 3 month livespan. Expired keys will be removed.

Documentation created in readme.md.